Get validateAndUpgradeLegacyGridFile.test.ts working again
