+++ src/debugFlags.ts
@@ -1,4 +1,4 @@
-const url = typeof window !== 'undefined' ? new URLSearchParams(window.location.search) : new URLSearchParams();
+const url = new URLSearchParams(window.location.search);
 
 // set this in .env (if set to false then all debug flags are turned off)
 export const debug = url.has('debug') || process.env.REACT_APP_DEBUG === '1' ? true : false;
